[DEV-1.6] Separated all UI scenes into their own class, organized variables, wrote javadocs for variables, wrote comments, and fixed bug where clients can still join a server even when the game has already begun
