+++ views/TetrisView.java
@@ -6,6 +6,7 @@ import javafx.beans.binding.BooleanBinding;
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.scene.effect.ColorAdjust;
+import model.TetrisApp;
 import model.TetrisModel;
 
 import javafx.animation.KeyFrame;
@@ -38,36 +39,26 @@ import java.util.ArrayList;
  */
 public class TetrisView {
 
+    // UI Variables
     public TetrisModel model; //reference to model
-    Stage stage;
-    Button singleplayerButton, chatButton, multiplayerButton, settingsButton, backButton; //buttons for functions
     public BorderPane borderPane;
+    Stage stage;
+    Button singleplayerButton, chatButton, multiplayerButton, settingsButton; //buttons for functions
     Canvas canvas;
     GraphicsContext gc; //the graphics context will be linked to the canvas
-    public ConnectView connectView;
-    public AnimationTimer timer;
-    private long lastUpdate = 0;
-    public Timeline timeline = new Timeline();
+
+    // Board Variables
     int pieceWidth = 20; //width of block on display
-    private double width; //height and width of canvas
-    private double height;
+    public double width; //height and width of canvas
+    public double height;
 
-    //settings variables
-    private ColorAdjust visualSettings;
-    private double brightness = 0;
-    private double saturation = 0;
-    private double contrast = 0;
-    private String backgroundColor = "black";
+    // Scene References;
+    public ConnectView connectView;
+    public SettingsView settingsView;
+    public GameView gameView;
 
-    // Variables for in-game controls. They check which buttons are pressed
-    BooleanProperty rotatePressed = new SimpleBooleanProperty();
-    BooleanProperty rightPressed = new SimpleBooleanProperty();
-    BooleanProperty leftPressed = new SimpleBooleanProperty();
-    BooleanProperty downPressed = new SimpleBooleanProperty();
-    BooleanProperty dropPressed = new SimpleBooleanProperty();
-    BooleanBinding anyPressed = downPressed.or(rightPressed).or(leftPressed).or(rotatePressed);
-    
-    private static TetrisView instance; // Instance reference for singleton
+    // Instance reference for singleton
+    private static TetrisView instance;
 
     /**
      * Constructor
@@ -104,7 +95,7 @@ public class TetrisView {
         this.height = this.model.getHeight()*pieceWidth + 2;
 
         borderPane = new BorderPane();
-        borderPane.setStyle("-fx-background-color: " + backgroundColor);
+        borderPane.setStyle("-fx-background-color: " + SettingsView.backgroundColor);
 
         //add buttons
         singleplayerButton = new Button("Singleplayer");
@@ -141,459 +132,25 @@ public class TetrisView {
         //Make sure to return the focus to the borderPane once you're done!
         singleplayerButton.setOnAction(e -> {
             //TO DO!
-            initSinglePlayerUI();
+            createSinglePlayerView();
             this.model.newGame();
             this.borderPane.requestFocus();
         });
 
         settingsButton.setOnAction(e -> {
-            initSettings();
+            createSettingsView();
             this.borderPane.requestFocus();
         });
 
-        visualSettings = new ColorAdjust();
-        visualSettings.setBrightness(brightness);
+        //visualSettings = new ColorAdjust();
+        //visualSettings.setBrightness(brightness);
 
         borderPane.setCenter(controls);
-        updateSettings();
+        SettingsView.updateSettings(borderPane);
 
         var scene = new Scene(borderPane, 800, 800);
         this.stage.setScene(scene);
         this.stage.show();
-    }
-
-    /**
-     * Initialize Settings interface
-     */
-    public void initSettings() {
-        backButton = new Button("Back");
-        backButton.setId("Settings");
-        backButton.setPrefSize(150, 50);
-        backButton.setFont(new Font(12));
-        backButton.setStyle("-fx-background-color: #17871b; -fx-text-fill: white;");
-
-        //brightness slider
-        Slider brightnessSlider = new Slider(0, 1, 0.5);
-        brightnessSlider.setShowTickLabels(true);
-        brightnessSlider.setStyle("-fx-control-inner-background: palegreen;");
-        //brightness label
-        Label brightnessLabel = new Label("Brightness");
-        brightnessLabel.setFont(new Font(20));
-        brightnessLabel.setTextFill(Color.WHITE);
-
-        //saturation slider
-        Slider saturationSlider = new Slider(0, 1, 0.5);
-        saturationSlider.setShowTickLabels(true);
-        saturationSlider.setStyle("-fx-control-inner-background: palegreen;");
-        //saturation label
-        Label saturationLabel = new Label("Saturation");
-        saturationLabel.setFont(new Font(20));
-        saturationLabel.setTextFill(Color.WHITE);
-
-        //contrast slider
-        Slider contrastSlider = new Slider(0, 1, 0.5);
-        contrastSlider.setShowTickLabels(true);
-        contrastSlider.setStyle("-fx-control-inner-background: palegreen;");
-        //contrast label
-        Label contrastLabel = new Label("Contrast");
-        contrastLabel.setFont(new Font(20));
-        contrastLabel.setTextFill(Color.WHITE);
-
-        //volume slider
-        Slider volumeSlider = new Slider(0, 100, 50);
-        volumeSlider.setShowTickLabels(true);
-        volumeSlider.setStyle("-fx-control-inner-background: palegreen;");
-        //volume label
-        Label volumeLabel = new Label("Volume");
-        volumeLabel.setFont(new Font(20));
-        volumeLabel.setTextFill(Color.WHITE);
-
-        //background color selector
-        ComboBox backGroundColor = new ComboBox();
-        backGroundColor.getItems().addAll("Red", "Blue", "Green", "Yellow", "Black", "White");
-        backGroundColor.setValue("Black");
-        //background color label
-        Label backGroundColorLabel = new Label("Background Color");
-        backGroundColorLabel.setFont(new Font(20));
-        backGroundColorLabel.setTextFill(Color.WHITE);
-        //background color confirm button
-        Button changeColorButton = new Button("Confirm Background Change");
-        changeColorButton.setId("Settings");
-        changeColorButton.setPrefSize(150, 50);
-        changeColorButton.setFont(new Font(12));
-        changeColorButton.setStyle("-fx-background-color: #17871b; -fx-text-fill: white;");
-
-        //hbox containing background color changing
-        HBox colorChange = new HBox(20, backGroundColorLabel, backGroundColor, changeColorButton);
-        colorChange.setPadding(new Insets(20, 20, 20, 20));
-        colorChange.setAlignment(Pos.CENTER);
-
-        //vbox containing all visual settings
-        VBox visualSettings = new VBox(20, backButton, brightnessLabel,
-                brightnessSlider, saturationLabel, saturationSlider, contrastLabel,
-                contrastSlider, volumeLabel, volumeSlider, colorChange);
-        visualSettings.setPadding(new Insets(20, 20, 20, 20));
-        visualSettings.setAlignment(Pos.CENTER);
-
-        //control settings label
-        Label controlSettingsLabel = new Label("Controls");
-        controlSettingsLabel.setFont(new Font(20));
-        controlSettingsLabel.setTextFill(Color.WHITE);
-
-        Button testButton = new Button("Test");
-        testButton.setId("Test");
-        testButton.setPrefSize(150, 50);
-        testButton.setFont(new Font(12));
-        testButton.setStyle("-fx-background-color: #17871b; -fx-text-fill: white;");
-
-        //vbox containing all control settings
-        VBox controlSettings = new VBox(20, controlSettingsLabel, testButton);
-        controlSettings.setPadding(new Insets(20, 20, 20, 20));
-        controlSettings.setAlignment(Pos.CENTER);
-
-        //hbox containing all settings
-        HBox settings = new HBox(50);
-        settings.setPadding(new Insets(20, 20, 20, 20));
-        settings.setAlignment(Pos.CENTER);
-        settings.getChildren().addAll(visualSettings, controlSettings);
-
-
-        backButton.setOnAction(e -> {
-            initUI();
-        });
-
-        brightnessSlider.setOnMouseReleased(e -> {
-            brightness = (brightnessSlider.getValue() - 0.5);
-            updateSettings();
-        });
-
-        saturationSlider.setOnMouseReleased(e -> {
-            saturation = (saturationSlider.getValue() - 0.5);
-            updateSettings();
-        });
-
-        contrastSlider.setOnMouseReleased(e -> {
-            contrast = (contrastSlider.getValue() - 0.5);
-            updateSettings();
-        });
-
-        volumeLabel.setOnMouseReleased(e -> {
-
-        });
-
-        changeColorButton.setOnAction(e -> {
-            backgroundColor = (String) backGroundColor.getValue();
-            borderPane.setStyle("-fx-background-color: " + backgroundColor);
-        });
-
-        borderPane = new BorderPane();
-        borderPane.setCenter(settings);
-        borderPane.setStyle("-fx-background-color: " + backgroundColor);
-        updateSettings();
-
-        var scene = new Scene(borderPane, 800, 800);
-        this.stage.setScene(scene);
-        this.stage.show();
-    }
-
-    /**
-     * Initialize Singleplayer Game Interface
-     */
-    public void initSinglePlayerUI() {
-        this.stage.setTitle("CSC207 Tetris");
-
-        borderPane = new BorderPane();
-        borderPane.setStyle("-fx-background-color: " + backgroundColor);
-
-        //add canvas
-        canvas = new Canvas(this.width, this.height);
-        canvas.setId("Canvas");
-        gc = canvas.getGraphicsContext2D();
-
-        final ToggleGroup addGarbageGroup = new ToggleGroup();
-
-        RadioButton addGarbageToggle = new RadioButton("Garbage");
-        addGarbageToggle.setToggleGroup(addGarbageGroup);
-        addGarbageToggle.setSelected(true);
-        addGarbageToggle.setUserData(Color.SALMON);
-        addGarbageToggle.setFont(new Font(16));
-        addGarbageToggle.setStyle("-fx-text-fill: #e8e6e3");
-
-        Slider addGarbageSpeed = new Slider(0, 100, 50);
-        addGarbageSpeed.setShowTickLabels(true);
-        addGarbageSpeed.setStyle("-fx-control-inner-background: palegreen;");
-
-        addGarbageGroup.selectedToggleProperty().addListener((observable, oldVal, newVal) -> swapGarbage(newVal));
-        addGarbageSpeed.setOnMouseReleased(e -> {
-            //TO DO
-            //double rateMultiplier = addGarbageSpeed.getValue() * 0.03;
-            //this.timeline.setRate(rateMultiplier);
-            //this.borderPane.requestFocus();
-        });
-
-        VBox botBox = new VBox(20, addGarbageSpeed);
-        botBox.setPadding(new Insets(20, 20, 20, 20));
-        botBox.setAlignment(Pos.TOP_CENTER);
-
-        VBox rightBox = new VBox(20, addGarbageToggle);
-        rightBox.setPadding(new Insets(20, 20, 20, 20));
-        botBox.setAlignment(Pos.TOP_CENTER);
-
-        // Detecting controls press
-        dropPressed.set(false);
-        borderPane.setOnKeyPressed(new EventHandler<KeyEvent>() {
-            @Override
-            public void handle(KeyEvent k) {
-                //TO DO
-                if (k.getCode() == KeyCode.SPACE) {
-                    rotatePressed.set(true);
-                    model.canPlace = false;
-                }
-                if (k.getCode() == KeyCode.S) {
-                    downPressed.set(true);
-                    model.canPlace = false;
-                }
-                if (k.getCode() == KeyCode.A) {
-                    leftPressed.set(true);
-                    model.canPlace = false;
-                }
-                if (k.getCode() == KeyCode.D) {
-                    rightPressed.set(true);
-                    model.canPlace = false;
-                }
-                if (k.getCode() == KeyCode.W) {
-                    if (!dropPressed.get()) {
-                        model.modelTick(TetrisModel.MoveType.DROP);
-                        paintBoard();
-                        dropPressed.set(true);
-                    }
-                }
-            }
-        });
-
-        // Timer for managing the block movement speed
-        timer = new AnimationTimer() {
-            @Override
-            public void handle(long now) {
-                if (now - lastUpdate >= 80_000_000) { // Prevent this loop from occurring more than once every 80 milliseconds
-                    if (rotatePressed.get()) {
-                        model.modelTick(TetrisModel.MoveType.ROTATE);
-                        paintBoard();
-                        rotatePressed.set(false);
-                    }
-                    if (downPressed.get()) {
-                        if (timeline.getStatus() != Animation.Status.RUNNING) {
-                            timeline = new Timeline(new KeyFrame(Duration.seconds(0.25), e -> {
-                                model.modelTick(TetrisModel.MoveType.DOWN);
-                                paintBoard();
-                            }));
-                            timeline.setCycleCount(Timeline.INDEFINITE);
-                            timeline.play();
-                        }else {
-                            timeline.setRate(timeline.getCurrentRate() + 0.25);
-                        }
-                    }
-                    if (rightPressed.get()) {
-                        model.modelTick(TetrisModel.MoveType.RIGHT);
-                        paintBoard();
-                    }
-                    if (leftPressed.get()) {
-                        model.modelTick(TetrisModel.MoveType.LEFT);
-                        paintBoard();
-                    }
-                    lastUpdate = now;
-                }
-            }
-        };
-
-        // Checking when the player releases their finger from a button
-        borderPane.setOnKeyReleased(new EventHandler<KeyEvent>() {
-            @Override
-            public void handle(KeyEvent k) {
-                //TO DO
-                if (k.getCode() == KeyCode.S) {
-                    downPressed.set(false);
-                }
-                if (k.getCode() == KeyCode.A) {
-                    leftPressed.set(false);
-                }
-                if (k.getCode() == KeyCode.D) {
-                    rightPressed.set(false);
-                }
-                if (k.getCode() == KeyCode.W) {
-                    dropPressed.set(false);
-                }
-            }
-        });
-
-        // Checking if the player pressed any button
-        anyPressed.addListener((obs, wasPressed, isNowPressed) -> {
-            if (isNowPressed) {
-                timer.start();
-            }else {
-                model.canPlace = true;
-                timer.stop();
-            }
-        });
-
-        borderPane.setCenter(canvas);
-        borderPane.setRight(rightBox);
-        borderPane.setBottom(botBox);
-        updateSettings();
-
-        var scene = new Scene(borderPane, 800, 800);
-        this.stage.setScene(scene);
-        this.stage.show();
-    }
-
-    /*
-    Updates settings according to the settings variables
-    * */
-    private void updateSettings(){
-        visualSettings.setBrightness(brightness);
-        visualSettings.setSaturation(saturation);
-        visualSettings.setContrast(contrast);
-        borderPane.setEffect(visualSettings);
-    }
-
-    /**
-     * Initialize Multiplayer Game Interface
-     */
-    public void initGameUI() {
-        this.stage.setTitle("CSC207 Tetris");
-
-        borderPane = new BorderPane();
-        borderPane.setStyle("-fx-background-color: " + backgroundColor);
-
-        //add canvas
-        canvas = new Canvas(this.width, this.height);
-        canvas.setId("Canvas");
-        gc = canvas.getGraphicsContext2D();
-
-        //labels
-        chatButton = new Button("Chat");
-        chatButton.setId("Chat");
-        chatButton.setPrefSize(150, 50);
-        chatButton.setFont(new Font(12));
-        chatButton.setStyle("-fx-background-color: #17871b; -fx-text-fill: white;");
-
-        HBox controls = new HBox(20, chatButton);
-        controls.setPadding(new Insets(20, 20, 20, 20));
-        controls.setAlignment(Pos.CENTER);
-
-        chatButton.setOnAction(e -> {
-            //TO DO!
-            this.createChatView();
-            this.borderPane.requestFocus();
-        });
-
-        // Detecting controls press
-        dropPressed.set(false);
-        borderPane.setOnKeyPressed(new EventHandler<KeyEvent>() {
-            @Override
-            public void handle(KeyEvent k) {
-                //TO DO
-                if (k.getCode() == KeyCode.SPACE) {
-                    rotatePressed.set(true);
-                    model.canPlace = false;
-                }
-                if (k.getCode() == KeyCode.S) {
-                    downPressed.set(true);
-                    model.canPlace = false;
-                }
-                if (k.getCode() == KeyCode.A) {
-                    leftPressed.set(true);
-                    model.canPlace = false;
-                }
-                if (k.getCode() == KeyCode.D) {
-                    rightPressed.set(true);
-                    model.canPlace = false;
-                }
-                if (k.getCode() == KeyCode.W) {
-                    if (!dropPressed.get()) {
-                        model.modelTick(TetrisModel.MoveType.DROP);
-                        paintBoard();
-                        dropPressed.set(true);
-                    }
-                }
-            }
-        });
-
-        // Timer for managing the block movement speed
-        timer = new AnimationTimer() {
-            @Override
-            public void handle(long now) {
-                if (now - lastUpdate >= 80_000_000) { // Prevent this loop from occurring more than once every 80 milliseconds
-                    if (rotatePressed.get()) {
-                        model.modelTick(TetrisModel.MoveType.ROTATE);
-                        paintBoard();
-                        rotatePressed.set(false);
-                    }
-                    if (downPressed.get()) {
-                        if (timeline.getStatus() != Animation.Status.RUNNING) {
-                            timeline = new Timeline(new KeyFrame(Duration.seconds(0.25), e -> {
-                                model.modelTick(TetrisModel.MoveType.DOWN);
-                                paintBoard();
-                            }));
-                            timeline.setCycleCount(Timeline.INDEFINITE);
-                            timeline.play();
-                        }else {
-                            timeline.setRate(timeline.getCurrentRate() + 0.25);
-                        }
-                    }
-                    if (rightPressed.get()) {
-                        model.modelTick(TetrisModel.MoveType.RIGHT);
-                        paintBoard();
-                    }
-                    if (leftPressed.get()) {
-                        model.modelTick(TetrisModel.MoveType.LEFT);
-                        paintBoard();
-                    }
-                    lastUpdate = now;
-                }
-            }
-        };
-
-        // Checking when the player releases their finger from a button
-        borderPane.setOnKeyReleased(new EventHandler<KeyEvent>() {
-            @Override
-            public void handle(KeyEvent k) {
-                //TO DO
-                if (k.getCode() == KeyCode.S) {
-                    downPressed.set(false);
-                }
-                if (k.getCode() == KeyCode.A) {
-                    leftPressed.set(false);
-                }
-                if (k.getCode() == KeyCode.D) {
-                    rightPressed.set(false);
-                }
-                if (k.getCode() == KeyCode.W) {
-                    dropPressed.set(false);
-                }
-            }
-        });
-
-        // Checking if the player pressed any button
-        anyPressed.addListener((obs, wasPressed, isNowPressed) -> {
-            if (isNowPressed) {
-                timer.start();
-            }else {
-                model.canPlace = true;
-                timer.stop();
-            }
-        });
-
-        borderPane.setTop(controls);
-        borderPane.setCenter(canvas);
-        updateSettings();
-
-        var scene = new Scene(borderPane, 800, 800);
-        this.stage.setScene(scene);
-        this.stage.show();
-    }
-
-    private void swapGarbage(Toggle newVal) {
     }
 
     /**
@@ -611,7 +168,6 @@ public class TetrisView {
     private final float dY() {
         return( ((float)(this.height-2)) / this.model.getBoard().getHeight() );
     }
-
 
     /**
      * Draw the board
@@ -656,10 +212,32 @@ public class TetrisView {
     }
 
     /**
-     * Create the view to chat with players in the lobby
+     * Create the settings view to modify the client's settings
      */
-    private void createChatView(){
-        // TODO: Change LoadView class into a UI for chatting
+    private void createSettingsView() {
+        settingsView = new SettingsView();
+    }
+
+    /**
+     * Create a single player game view
+     */
+    private void createSinglePlayerView() {
+        //add canvas
+        canvas = new Canvas(this.width, this.height);
+        canvas.setId("Canvas");
+        gc = canvas.getGraphicsContext2D();
+        gameView = new SinglePlayerView();
+    }
+
+    /**
+     * Create a multiplayer game view
+     */
+    public void createMultiplayerView() {
+        //add canvas
+        canvas = new Canvas(this.width, this.height);
+        canvas.setId("Canvas");
+        gc = canvas.getGraphicsContext2D();
+        gameView = new MultiplayerView();
     }
 
     /**
